location: Move to backend/location
